requirements.txt
groq==0.9.0
numpy==2.0.0
pydub==0.25.1
python-dotenv==1.0.1
scipy==1.14.0
TTS==0.22.0
TTS==0.22.0
websockets==12.0
